Added lines to view_results.php to see entire contents of results directory, to see if .zip files are being created correctly.

+++ download_results.php
@@ -53,8 +53,9 @@
     header('Content-Length: ' . filesize($archive_path));
     // ob_clean();
     flush();
-    readfile($archive_path);
-    unlink($archive_path);
+    if(readfile($archive_path)) {
+      unlink($archive_path);
+    }
 
     // Remove files
